feat: sending backend image to html
