--- main.py
@@ -1,8 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-from views import *
-
-if __name__ == "__main__":
-    app.run()
